fix: ccp: report the orders a reconnect does not explain
An order working when the connection drops is marked Uncertain, and that was
the end of it. The engine stopped believing the status and told nobody, so the
API layer went on reporting the pre-disconnect status for the whole outage.

Two things were already arriving and being discarded.

A fractional order reports what is actually outstanding, and reports it
exactly. It tracks its quantity as zero — the decimal it was submitted with
lives only in the enriched record — so subtracting from that reported nothing
outstanding on an order that is entirely outstanding. The quantity fields on
the update were only integers because the struct declared them that way, not
because the value itself is one — both dispatchers were already casting them to
floating point on the way to the callback — so the fix carries the real
fraction through end to end rather than rounding it into a field that never
needed to be a whole number: half a share outstanding now reports 0.5, and a
partly filled fractional order reports exactly what is left, not a number
rounded up to the next whole share. A disconnect landing before the gateway has
sent even the first word on a freshly submitted fractional order finds no
enriched record either, so there is no quantity to give at all; the update
carries ibapi's own sentinel for a value that was never told to it, rather than
a guessed number standing in for one.

The flip itself was silent: `mark_orders_uncertain` mutated the book and said
nothing. The disconnect now publishes an order status update for each order it
withdraws, carrying the filled and remaining quantities and the identifiers a
caller keys on — the broker id and the bracket link, read back from the
enriched cache rather than zeroed.

The open-order snapshot then reasserted the old answer anyway. It unions the
client's own records with the cached view, and the cached view still said the
order was working, so for the whole outage a caller asking which orders were
open was told the opposite of what the callback had just told them. The union
exists to add orders this client has no record of, such as one placed in an
earlier session; it now skips the ones whose status this client has withdrawn,
and only those, so a genuinely newer cached status still merges.

That withdrawn set is only as complete as this client's own local records,
though, and an order recovered from an earlier session has none — this client
never submitted it, so nothing had created a local entry for it in the first
place. The status update meant to withdraw it found nothing to update, did
nothing, and the stale cached record kept coming back, for precisely the order
this same fix's union exists to admit. It now creates that entry the moment one
is needed, seeded from the same enriched cache the snapshot itself reads, so an
order this client is hearing about for the first time is both withdrawn
correctly and, should that first word about it be something other than a
withdrawal, not handed back with an empty contract either.

What this deliberately does not do is decide what became of an order the
recovery push never mentions. Absence does not identify an outcome. It cannot
separate a cancellation from a rejection. An order carrying a decimal quantity
holds zero in the integer quantity field, so a completeness test reads it as
filled when nothing filled. And an execution for the order can still arrive
after the push ends, which it cannot book if the order has been retired from
the book to record a verdict — the fill and the position move are lost, which
is worse than the state being unknown. The order stays Uncertain and stays in
the book, where any real message supersedes it, and the caller is told the
status is unknown rather than told something that may be wrong.

Deciding those outcomes needs a positive statement from the gateway rather than
an inference from silence — a status request on reconnect, or a capture
establishing what the push guarantees. Refs #251, which also carries a
`PendingSubmit` ack deadline and the HMDS warm-up re-arm.

No new wire traffic: this consumes what the gateway already sends.

+++ src/python/compat/client/test_helpers.rs
@@ -103,7 +103,7 @@ impl EClient {
     #[doc(hidden)]
     fn _test_push_order_update(
         &self, order_id: u64, instrument: u32, status: &str,
-        filled_qty: i64, remaining_qty: i64,
+        filled_qty: f64, remaining_qty: f64,
     ) -> PyResult<()> {
         let shared = self.shared_state()?;
         let st = match status {
